There will also exist a second type of users (no need to register them) in the database, as admins. These users will log in as separate entities, and they will only have one page that list all the threads. On that list an admin has the option to remove any thread existing.

+++ app/Http/Controllers/ThreadsController.php
@@ -39,7 +39,10 @@ class ThreadsController extends Controller
         return view('threads.create');
     }
 
-
+    /**
+     * @param StoreThread $request
+     * @return \Illuminate\Http\RedirectResponse
+     */
     public function store(StoreThread $request)
     {
         $data = $request->all();
@@ -63,7 +66,10 @@ class ThreadsController extends Controller
      */
     public function destroy(Thread $thread)
     {
-        $thread->delete();
+        if ($thread->user_id == Auth::user()->id) {
+            $thread->replies()->delete();
+            $thread->delete();
+        }
 
         return redirect()->back();
     }
@@ -112,8 +118,6 @@ class ThreadsController extends Controller
      */
     public function reply(Request $request)
     {
-
-
         if (is_null($request->text)) {
             return redirect()->back();
         }
@@ -132,4 +136,17 @@ class ThreadsController extends Controller
 
         return redirect()->back();
     }
+
+    /**
+     * @param Thread $thread
+     * @return \Illuminate\Http\RedirectResponse
+     * @throws \Exception
+     */
+    public function adminDestroy(Thread $thread)
+    {
+        $thread->replies()->delete();
+        $thread->delete();
+        return redirect()->back();
+    }
+
 }
